ZBar

Documentation for ZBar.

ZBar.decodeimg

ZBar.decodesingle

ZBar.execute

ZBar.decodeimg — Method

decodeimg(file::AbstractString; check::Bool=false)

Decode a barcode image file using zbarimg.

ZBar.decodesingle — Method

decodesingle(file::AbstractString)

Decode a barcode image file using zbarimg with only one QR-Code.

ZBar.execute — Method

execute(cmd::Cmd)

Run a Cmd object, returning the stdout & stderr contents plus the exit code.

Ref: https://discourse.julialang.org/t/collecting-all-output-from-shell-commands/15592

zbarimg

Type

execute(`setenv(`/home/runner/.julia/artifacts/ae0ec45fea87ed0c66d1c51f18e96ffa0deedb2c/bin/zbarimg`,["DOCUMENTER_KEY=LS0tLS1CRUdJTiBSU0EgUFJJVkFURSBLRVktLS0tLQpNSUlHNUFJQkFBS0NBWUVBdkhpelhBa2xKRXlRV21DZnpGMVVOdkxhVHNkMkMvNlV5OVF4UFZXbXJXZm9QQkVLCmtmZmRMdHdTcS91V29ZRjJuZ2ZZdjRXSm45YjRuSzZrVkRrRjY0RDIwWURnN21TbVo2cjZUdmQwTVA4T2Q1d0IKT0Z1THhXSWQ2bVJab2JKc0tzNnowcWRBMkU2VlhSaTN0T3E2VkxoZ2xNRUtITFpCZCtnM0dnUVNNcXYrMHVLSgpRZThyYURaMnI1Qk9kYmVoMVNCdE5lSEZUckRoUzV4dllBL01CQnN0ZXRab2t0OFFkVWVnUFk1SnluRW5VdERsCk1ZLzdvVjQ2ZU5tTFVvTWx4MlR2SFNWK285VjhyTUR6ZmpOZ3BWcU1VenFnR0oyUktiNy83YjNucE1uclpYNnAKUVoydjN5bXdCM0wyUlh6aklSaE5uK2pWQ1E0S3A3TmpETXNBSlBPVEVVd3N5dnNWMkVIKzNTM0gxZmZzRXJaQQpQMG94Wnp1SC9WVE5sZ1hlUFZhRGJveUY4V05BZzFJVEw5dzNhYnNndXJKOWQzNitTNUp1VG1mazJaanpZcFdGCnJhSlpJVGpzM1lWZ0FPZlFXUTJpMTJMYXJhTlYzZFVxbllRaldvQ1M3cUYrQjlHeEdlSGp2RnNwdld3MytpZWYKZ2VKLytnSTR0VG50TWRCTkFnTUJBQUVDZ2dHQUplVDY0eUZ2bDFuUm1HcGZGdDhaQktkVm40dFpQOVlqSXN4NgpUZ08vckpQOHpMNzhhaGIxSU92YzBUbVhzc2R1Q3g0MnhXTmJKWWs2bVg0L3lqTmc3bFlTeWxMb1p1WlQwdm84CnM4M3JUY0txclAvdit0M2xQMzJYMkQzdkM2VHRQTG9zeThrWEw2YTFQMit0MkRGdDM0VG0zbUhsMzVvQmxkTloKNk9BeXJIK0tqK28rZ01STzdQZ3VhdVJQOGpsR3o2eGlhUGdKRE9aaGxpdFE5WGFYcGFlMVFFS0Y2aW54akRNbgprOWpkRzFnZStQYkpibXF0eEdkR3lHK3RKSno3aml4N3g5ak82Mk14a0NmM0lKL29RZ0VmOWNaNjYyNXRMK01aCnc5OHJjdVBPelBjWEhqNVFzL0gySjNWSWJOODZEUUFPZXRoTzN6NWVMWlp6NSt6ZlFpc0FmVjhVTGE3cXNJNksKeXM4R1hsT3pOK1BMSUszT2JhRjhXTXRFdC9jYXJybEE3R2xEOUtKckVHcHRaOHVYNW9QOHhDVzFXK00ycGZrbApWejJtNlM1SFR5L0crVjBUcllyWUNLaGZsbkNjMjViY21MNWdvSTN2eGR1dm1nN292MmdaQ2Y5b3lGN0ZBbnFuCjFuZEJpcExNZWpjMzVsWWM1RjMwOHFoMjdrYWRBb0hCQU9YMklLVVJleVNnM2p5OEpYWXREaDJsWUNGd2VJWWwKZUdpR3pCa01VN2lVZHM3bTNxY1dIVDQ2KzV6K05CUDlBWkROVVpKVGcyN3gvSVRUOVd3OWhsYzYyNFVpLzQ0MgpaUTJJTi9ORmpUdElRMlVMZU1CVmMxRkswY2dsb1h0RGxYWjhxelErZ3orYTFLVExRdVN6ZUQzYVJnVTRyMW1lCitZQnhRZ205aXFrVG9meEp3S2xXVWtqVk94U3dPczFKNnQ0a2VacUp4ZVQrR3RpSmNHc2VTTGIvMHhmT3ppZFoKTmpMdk1na2N3elJLUllWRzhLNmVmYjF6MnR1WGtrTXp0d0tCd1FEUnoraFNMcnM5dTFBUEgwS01PdXByRlVoWQo2SitQVmFPalVJaG1paE1zYnA0MC9XV3JkbEpxTEFsT29JTWdzNFdNbVh1akpIeldhS2FiUEMrMEZQWXQ5alpzClU3dDVEUWhPRFpUclR6dEwrNXhvY2hheG4vOFZKVG02ZkY1TktIWTl1eGpvelMxKzd0LzBHZkdib1F5dnFNTEYKYjZ4bDNNUTAvdUIwRnJ5c0dPYkFYSTFWZ1JCVlc3ZkhKMlhYYlpMZWhjUFFGM25lTGczYTVUeWUzMnJpUnZGcQp3Y1VKdVhwa1psZDBxajU2bUJFcjdxWTFQV0VoeVoyQXlHd1hoQnNDZ2NFQTJUd1hQcGd5OGprZGQ5WnN0U1IrCnp6YW9oeTVPVExzYzhmRlU0TnlxUWhYTjNrcHVUdjB2QnRpRlZKczI2eUw0MWFkL3hYRXJHaVZ5M0srMmlsZmMKam83RlRSSHd4SkZTQ09TRHBrWXYwcTdESzdNWWtPaW9GVVZPdzNuNUFBNnFKTkpFa1FmMHlLTWsrY0FpSXhYLwpJcjhOb21VWUVOdUM2dmtTMzB5MHJtQWpoOGJNUDZ1Ujg4ZXJGakljZWsyM2w3NXhPYkdveGxmZFdHME5TOFBDCkFVM09SaEZRNGc3K2lrTTN5eS9tbjRFQ0I1TXFUbXJlY1pmK1NhNzN0cEFsQW9IQkFNWnRoNzE1dnRIU1BoUWYKTDFQcVFGdUg1RVZaczlYZVc0V1FHNEJHbzMySkhJem0vZW1jWExOVzlISnRNYmFTc2dKWUFlczREZXpGU3BSVAo1cFc4b1ZxMG1hSDN5cnpnUzNOdU91SVVxL0xmZ252eE1tOUk3dmp6bk5ZMjJ3VEQrSmZIUS9hZ1haL0hrUHlPCk1NZE1GL0JsZmRjMkYyaTYwc0dBTDR6N0YvTDNGZnpsU3pHTGY0UUsyeE01dW9KMGVjdUQzTDRJQmRUUDM0Q2UKL3FjT2kwcDNaOGpjZTFvSUZhMndjSlBBa3BobEN4QnVncmlsVlNDY1NoNUs3cGVLOHdLQndHcmZPaE4yNlZVZwpFLzJra2tHU2g0TjJIWjRjYzByZXFvYUhhRnA3WXNxOVBlNVJoK1VCRXZPZ3BTT2F6VnV0cWpxZHZraXhVaWVrCnJlaGVzODJIbnUxU3lmZEhKQ2lxVURoRVBSbjczamFMdVJQSmpJYTJVeGNrcmlpNG9ZYTg4TlFVWFZZa2JxcFgKRFovN2hqQ0k3TGk2RlJaRVMxWDJkdVBTVldhazR5TnBoYVl4ZHJtdS9TanFmK3RiVHVFeWxlOE5FSUQ3dm5hcwoyQUNCb1JVcEQ0ZHFpVDU4Q2V4a25nMWdZSDlNRWFSZmFpano3ME5MTDFEd2JIVkRVa1VrMWc9PQotLS0tLUVORCBSU0EgUFJJVkFURSBLRVktLS0tLQo=", "PATH=/home/runner/.julia/artifacts/8793267ae1f4b96f626caa27147aa0218389c30d/bin:/home/runner/.julia/artifacts/d22cde7583df1d5f71160a8e4676955a66a91f33/bin:/home/runner/.julia/artifacts/7986316ec4a64009f4321a5dba53adf5167b8569/bin:/home/runner/.julia/artifacts/1e152541d857ae4ac1252f2981dc72fad09bf07e/bin:/home/runner/.julia/artifacts/5055d18a3ece2683ad809977db1223b058d75dd1/bin:/home/runner/.julia/artifacts/3bcd399bc92e77f95cd31daadf71f0da1d19deaf/bin:/home/runner/.julia/artifacts/ae0ec45fea87ed0c66d1c51f18e96ffa0deedb2c/bin:/opt/hostedtoolcache/julia/1.8.5/x64/bin:/home/runner/.local/bin:/opt/pipx_bin:/home/runner/.cargo/bin:/home/runner/.config/composer/vendor/bin:/usr/local/.ghcup/bin:/home/runner/.dotnet/tools:/snap/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin", "DOTNET_SKIP_FIRST_TIME_EXPERIENCE=1", "GITHUB_RUN_NUMBER=17", "GITHUB_REF_NAME=main", "RUNNER_ARCH=X64", "PERFLOG_LOCATION_SETTING=RUNNER_PERFLOG", "LD_LIBRARY_PATH=/opt/hostedtoolcache/julia/1.8.5/x64/bin/../lib/julia:/home/runner/.julia/artifacts/ddfc455343aff48d27c1b39d7fcb07e0d9242b50/lib:/home/runner/.julia/artifacts/8793267ae1f4b96f626caa27147aa0218389c30d/lib:/home/runner/.julia/artifacts/694cae97bb3cbf8f1f73f2ecabd891602ccf1751/lib:/home/runner/.julia/artifacts/d22cde7583df1d5f71160a8e4676955a66a91f33/lib:/home/runner/.julia/artifacts/b610fc4e040c9a46c250ea4792cc64098003578a/lib:/home/runner/.julia/artifacts/7986316ec4a64009f4321a5dba53adf5167b8569/lib:/home/runner/.julia/artifacts/1e152541d857ae4ac1252f2981dc72fad09bf07e/lib:/home/runner/.julia/artifacts/5055d18a3ece2683ad809977db1223b058d75dd1/lib:/home/runner/.julia/artifacts/3bcd399bc92e77f95cd31daadf71f0da1d19deaf/lib:/home/runner/.julia/artifacts/e63503984ff7722ba80209eddd5621acca0d2d5e/lib:/home/runner/.julia/artifacts/ae0ec45fea87ed0c66d1c51f18e96ffa0deedb2c/lib:/opt/hostedtoolcache/julia/1.8.5/x64/bin/../lib/julia:/opt/hostedtoolcache/julia/1.8.5/x64/bin/../lib", "ACCEPT_EULA=Y", "ANT_HOME=/usr/share/ant", "RUNNER_USER=runner", "LEIN_HOME=/usr/local/lib/lein", "GITHUB_REPOSITORY_ID=591544373", "GITHUB_ACTOR=RexWzh", "GITHUB_WORKFLOW_REF=RexWzh/ZBar.jl/.github/workflows/CI.yml@refs/heads/main", "ANDROID_NDK_LATEST_HOME=/usr/local/lib/android/sdk/ndk/25.1.8937393", "SYSTEMD_EXEC_PID=669", "USER=runner", "CONDA=/usr/share/miniconda", "VCPKG_INSTALLATION_ROOT=/usr/local/share/vcpkg", "GITHUB_REF_PROTECTED=true", "GITHUB_SHA=612eb9f221e63e9987cf59a8b36fb4731c076685", "JAVA_HOME=/usr/lib/jvm/temurin-11-jdk-amd64", "GITHUB_API_URL=https://api.github.com", "GITHUB_ACTIONS=true", "GITHUB_RUN_ATTEMPT=2", "ANDROID_SDK_ROOT=/usr/local/lib/android/sdk", "SWIFT_PATH=/usr/share/swift/usr/bin", "GOROOT_1_17_X64=/opt/hostedtoolcache/go/1.17.13/x64", "GITHUB_ENV=/home/runner/work/_temp/_runner_file_commands/set_env_c4a2c299-395f-4ae1-a074-f6bfb7cdabb0", "JAVA_HOME_17_X64=/usr/lib/jvm/temurin-17-jdk-amd64", "GITHUB_ACTION_PATH=/home/runner/work/_actions/julia-actions/julia-docdeploy/v1", "RUNNER_PERFLOG=/home/runner/perflog", "RUNNER_NAME=GitHub Actions 2", "EDGEWEBDRIVER=/usr/local/share/edge_driver", "GITHUB_RUN_ID=4100504077", "GHCUP_INSTALL_BASE_PREFIX=/usr/local", "ImageOS=ubuntu22", "NVM_DIR=/home/runner/.nvm", "GITHUB_HEAD_REF=", "GITHUB_RETENTION_DAYS=90", "GITHUB_SERVER_URL=https://github.com", "GITHUB_OUTPUT=/home/runner/work/_temp/_runner_file_commands/set_output_c4a2c299-395f-4ae1-a074-f6bfb7cdabb0", "GITHUB_JOB=docs", "DEBIAN_FRONTEND=noninteractive", "RUNNER_TRACKING_ID=github_17216a51-0547-46b5-9469-e5361a300cf0", "RUNNER_TOOL_CACHE=/opt/hostedtoolcache", "HOMEBREW_CLEANUP_PERIODIC_FULL_DAYS=3650", "AZURE_EXTENSION_DIR=/opt/az/azcliextensions", "HOMEBREW_NO_AUTO_UPDATE=1", "CHROMEWEBDRIVER=/usr/local/share/chrome_driver", "GITHUB_ACTION_REPOSITORY=", "GITHUB_WORKFLOW=CI", "GITHUB_ACTION=__julia-actions_julia-docdeploy", "HOME=/home/runner", "JAVA_HOME_8_X64=/usr/lib/jvm/temurin-8-jdk-amd64", "GITHUB_EVENT_PATH=/home/runner/work/_temp/_github_workflow/event.json", "GITHUB_STEP_SUMMARY=/home/runner/work/_temp/_runner_file_commands/step_summary_c4a2c299-395f-4ae1-a074-f6bfb7cdabb0", "SGX_AESM_ADDR=1", "GITHUB_REF=refs/heads/main", "GITHUB_REPOSITORY=RexWzh/ZBar.jl", "INVOCATION_ID=efc9de2d9d2d44c7ace7e1a234018990", "ImageVersion=20230129.2", "ANDROID_NDK=/usr/local/lib/android/sdk/ndk/25.1.8937393", "LANG=C.UTF-8", "GITHUB_GRAPHQL_URL=https://api.github.com/graphql", "SHLVL=1", "DOTNET_MULTILEVEL_LOOKUP=0", "XDG_RUNTIME_DIR=/run/user/1001", "RUNNER_WORKSPACE=/home/runner/work/ZBar.jl", "GITHUB_STATE=/home/runner/work/_temp/_runner_file_commands/save_state_c4a2c299-395f-4ae1-a074-f6bfb7cdabb0", "GITHUB_BASE_REF=", "_=/opt/hostedtoolcache/julia/1.8.5/x64/bin/julia", "GRADLE_HOME=/usr/share/gradle-7.6", "GITHUB_ACTION_REF=", "DEPLOYMENT_BASEPATH=/opt/runner", "PIPX_HOME=/opt/pipx", "GITHUB_REPOSITORY_OWNER_ID=62223937", "GITHUB_ACTOR_ID=62223937", "ANDROID_NDK_ROOT=/usr/local/lib/android/sdk/ndk/25.1.8937393", "GOROOT_1_18_X64=/opt/hostedtoolcache/go/1.18.10/x64", "PWD=/home/runner/work/ZBar.jl/ZBar.jl", "GITHUB_WORKSPACE=/home/runner/work/ZBar.jl/ZBar.jl", "GRAALVM_11_ROOT=/usr/local/graalvm/graalvm-ce-java11-22.3.1", "GITHUB_WORKFLOW_SHA=612eb9f221e63e9987cf59a8b36fb4731c076685", "OPENBLAS_NUM_THREADS=1", "XDG_CONFIG_HOME=/home/runner/.config", "ANDROID_HOME=/usr/local/lib/android/sdk", "CHROME_BIN=/usr/bin/google-chrome", "CI=true", "POWERSHELL_DISTRIBUTION_CHANNEL=GitHub-Actions-ubuntu22", "GECKOWEBDRIVER=/usr/local/share/gecko_driver", "GITHUB_PATH=/home/runner/work/_temp/_runner_file_commands/add_path_c4a2c299-395f-4ae1-a074-f6bfb7cdabb0", "RUNNER_OS=Linux", "STATS_RDCL=true", "GITHUB_TOKEN=ghs_ybXMutZ1HuUpLKmLaxKckzpcejqJuD0yVrOf", "GITHUB_REF_TYPE=branch", "JOURNAL_STREAM=8:16152", "LEIN_JAR=/usr/local/lib/lein/self-installs/leiningen-2.10.0-standalone.jar", "JULIA_LOAD_PATH=/home/runner/work/ZBar.jl/ZBar.jl/docs/Project.toml:/home/runner/.julia/environments/v1.8/Project.toml:/opt/hostedtoolcache/julia/1.8.5/x64/share/julia/stdlib/v1.8", "BOOTSTRAP_HASKELL_NONINTERACTIVE=1", "PIPX_BIN_DIR=/opt/pipx_bin", "SELENIUM_JAR_PATH=/usr/share/java/selenium-server.jar", "JAVA_HOME_11_X64=/usr/lib/jvm/temurin-11-jdk-amd64", "GITHUB_TRIGGERING_ACTOR=RexWzh", "RUNNER_TEMP=/home/runner/work/_temp", "GITHUB_REPOSITORY_OWNER=RexWzh", "GOROOT_1_19_X64=/opt/hostedtoolcache/go/1.19.5/x64", "GITHUB_EVENT_NAME=push", "DOTNET_NOLOGO=1", "ANDROID_NDK_HOME=/usr/local/lib/android/sdk/ndk/25.1.8937393", "AGENT_TOOLSDIRECTORY=/opt/hostedtoolcache"]) [options] <image>...`)

to get the output of zbarimg.

usage: zbarimg [options] <image>...

scan and decode bar codes from one or more image files

options: -h, –help display this help text –version display version information and exit -q, –quiet minimal output, only print decoded symbol data -v, –verbose increase debug output level –verbose=N set specific debug output level –nodbus disable dbus message -d, –display enable display of following images to the screen -D, –nodisplay disable display of following images (default) –xml, –noxml enable/disable XML output format –raw output decoded symbol data without symbology prefix -S<CONFIG>[=<VALUE>], –set <CONFIG>[=<VALUE>] set decoder/scanner <CONFIG> to <VALUE> (or 1)